% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.ctmc.R
\name{discrete.ctmc}
\alias{discrete.ctmc}
\title{Discretize a continuous-time Markov chain (CTMC) trajectory}
\usage{
discrete.ctmc(ctmc.times, ctmc.states, obs.times)
}
\arguments{
\item{ctmc.times}{Transition times of the CTMC}

\item{ctmc.states}{States of the CTMC at corresponding transition times.}

\item{obs.times.}{Times at which observations are desired.}
}
\value{
A data frame containing the oberved times and corresponding states.
}
\description{
This function discretizes a CTMC trajectory given the transition times and states.
}
