% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_death.R
\name{sim_cancer_death}
\alias{sim_cancer_death}
\title{Simulate a time of cancer death for a specific cancer in a specific stage based on provided survival distributions.
Uses the inverse CDF method.}
\usage{
sim_cancer_death(
  age_clin_dx,
  stage_dx,
  cancer_type,
  cancer_survival_dist,
  model
)
}
\arguments{
\item{the_stage}{stage at clinical diagnosis}

\item{the_cancer_site}{cancer site (one of "Anus","Breast","Bladder","Colorectal","Esophagus","Headandneck","Gastric","Liver" ,"Lung","Pancreas", "Prostate", "Renal", "Ovary", "Uterine")}

\item{the_sex}{Male or Female}

\item{the_model_type}{Weibull or  Loglogistic}

\item{cancer_survival_dist:}{a data frame with columns: surv, time, site, stage, sex, model_type}
}
\description{
Simulate a time of cancer death for a specific cancer in a specific stage based on provided survival distributions.
Uses the inverse CDF method.
}
