% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_universe.R
\name{sim_individual_MCED}
\alias{sim_individual_MCED}
\title{Simulate a single individual with and without Multicancer Early Detection (MCED) screening}
\usage{
sim_individual_MCED(
  ID,
  cancer_sites,
  rates_list,
  test_performance,
  MCED_specificity,
  other_cause_death_dist,
  starting_age,
  num_screens,
  screen_interval,
  end_time,
  sex,
  surv_param_table
)
}
\arguments{
\item{ID}{Numeric ID for the individual.}

\item{cancer_sites}{Vector of selected cancer sites.}

\item{rates_list}{List of transition rate matrices for each cancer site.}

\item{test_performance}{A list with early_sens, late_sens, and specificities for the test.}

\item{MCED_specificity}{Overall specificity for the MCED test (not specific to cancer site)}

\item{other_cause_death_dist}{A table representing other-cause mortality.}

\item{starting_age}{Starting age for simulation.}

\item{num_screens}{Number of screening rounds.}

\item{screen_interval}{Time interval between screenings.}

\item{end_time}{Ending time (age) of simulation.}

\item{sex}{"Male" or "Female".}

\item{surv_param_table}{Survival parameters table (for simulating cancer death).}
}
\value{
A data frame with simulation results for the individual, including cancer events, screening detection, and death info.
}
\description{
This function simulates cancer outcomes in a single individual with a given starting age using a "parallel universe" approach.
That is, cancer outcomes are simulated with and without MCED screening.  The natural history (i.e., the times of cancer onset and clinical diagnosis) for 
is the same in both screening and no-screening scenarios.  
The user specifies cancer sites in the MCED screening test.  The user also provides sensitivity of the tests for early and late-stage disease, where early refers to AJCC 7 stages I-II and late, III-IV, except for
pancreas cancer, where early is stage I and late, II-IV. 
Natural history models for each cancer site are specified based on a list of transition rate matrices for each site.  
Cancer-specific mortality is specified based on a data frame providing the parametric survival parameters for each cancer site.  Built-in tables are included in the "param_table"
dataframe that loads with the built-in "parametric_surv_fits" data sets. 
Other-cause mortality is based on other-cause life tables that can be extracted with the function "make_othercause_death_table."
}
\details{
The function tracks only the first cancer diagnosis based on pre-clinical onset. 
Cancer-specific mortality in the screen arm assumes a stage-shift benefit of screening. That is, individuals
who are diagnosed in early stage under screening but who would have been diagnosed in late stage clinically, are assume to remain in early stage post lead-time. 
For both screened and unscreened individuals, cancer mortality is projected from the point of clinical diagnosis (i.e., post lead-time) to prevent
lead-time bias.
}
\examples{
result <- sim_individual_MCED(
  ID = 1,
  cancer_sites = c("Lung", "Colorectal"),
  rates_list = example_rates_list,
  test_performance = example_test_perf,
  other_cause_death_dist = example_death_dist,
  starting_age = 50,
  num_screens = 5,
  screen_interval = 1,
  end_time = 90,
  sex = "Male",
  surv_param_table = example_surv_table
)
}
