% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_universe.R
\name{sim_multiple_individuals_MCED_parallel_universe}
\alias{sim_multiple_individuals_MCED_parallel_universe}
\title{Simulate Multiple Individuals Under a Parallel Universe MCED Setting}
\usage{
sim_multiple_individuals_MCED_parallel_universe(
  cancer_sites,
  LMST_vec,
  OMST_vec,
  test_performance_dataframe,
  MCED_specificity,
  starting_age,
  ending_age,
  num_screens,
  screen_interval,
  num_males,
  num_females,
  all_rates_male,
  all_rates_female,
  all_meta_data_female,
  all_meta_data_male,
  cdc_data,
  hmd_data,
  MCED_cdc,
  surv_param_table
)
}
\arguments{
\item{cancer_sites}{Vector of cancer sites.}

\item{LMST_vec}{Vector of late mean sojourn times.}

\item{OMST_vec}{Vector of overall mean sojourn times.}

\item{test_performance_dataframe}{Data frame with test sensitivity/specificity info.}

\item{MCED_specificity}{Overall specificity for the MCED test (not specific to cancer site)}

\item{starting_age}{Starting age for simulation.}

\item{ending_age}{Ending age for simulation.}

\item{num_screens}{Number of screening rounds.}

\item{screen_interval}{Interval between screening rounds.}

\item{num_males}{Number of male individuals to simulate.}

\item{num_females}{Number of female individuals to simulate.}

\item{all_rates_male}{List of transition matrices for males.}

\item{all_rates_female}{List of transition matrices for females.}

\item{all_meta_data_female}{Metadata for female cancer sites.}

\item{all_meta_data_male}{Metadata for male cancer sites.}

\item{cdc_data}{CDC mortality data.}

\item{hmd_data}{Human Mortality Database data.}

\item{MCED_cdc}{CDC data for MCED.}

\item{surv_param_table}{Survival parameters table.}
}
\value{
A data frame with combined simulated results for all individuals.
}
\description{
Simulates multiple individuals with cancer onset, screening detection, and mortality (cancer or other cause).
}
\examples{
sim_data <- sim_multiple_individuals_MCED_parallel_universe(
  cancer_sites = c("Lung", "Colorectal"),
  LMST_vec = c(2, 3),
  OMST_vec = c(5, 6),
  test_performance_dataframe = test_perf_df,
  starting_age = 50,
  ending_age = 90,
  num_screens = 5,
  screen_interval = 1,
  num_males = 100,
  num_females = 100,
  all_rates_male = male_rates,
  all_rates_female = female_rates,
  all_meta_data_female = female_meta,
  all_meta_data_male = male_meta,
  cdc_data = cdc,
  hmd_data = hmd,
  MCED_cdc = mced_cdc,
  surv_param_table = surv_params
)

}
